Move create_train_model, create_eval_model, create_infer_model to model_helper.py plus some clean-ups.
PiperOrigin-RevId: 166518125
