+++ nmt/inference.py
@@ -22,67 +22,16 @@ import time
 
 import tensorflow as tf
 
-from tensorflow.python.ops import lookup_ops
-
 from . import attention_model
 from . import gnmt_model
 from . import model as nmt_model
 from . import model_helper
-from .utils import iterator_utils
 from .utils import misc_utils as utils
 from .utils import nmt_utils
-from .utils import vocab_utils
 
-__all__ = ["create_infer_model", "load_data", "inference",
+__all__ = ["load_data", "inference",
            "single_worker_inference", "multi_worker_inference"]
 
-
-class InferModel(
-    collections.namedtuple("InferModel",
-                           ("graph", "model", "src_placeholder",
-                            "batch_size_placeholder", "iterator"))):
-  pass
-
-
-def create_infer_model(model_creator, hparams, scope=None, single_cell_fn=None):
-  """Create inference model."""
-  graph = tf.Graph()
-  src_vocab_file = hparams.src_vocab_file
-  tgt_vocab_file = hparams.tgt_vocab_file
-
-  with graph.as_default():
-    src_vocab_table, tgt_vocab_table = vocab_utils.create_vocab_tables(
-        src_vocab_file, tgt_vocab_file, hparams.share_vocab)
-    reverse_tgt_vocab_table = lookup_ops.index_to_string_table_from_file(
-        tgt_vocab_file, default_value=vocab_utils.UNK)
-
-    src_placeholder = tf.placeholder(shape=[None], dtype=tf.string)
-    batch_size_placeholder = tf.placeholder(shape=[], dtype=tf.int64)
-
-    src_dataset = tf.contrib.data.Dataset.from_tensor_slices(
-        src_placeholder)
-    iterator = iterator_utils.get_infer_iterator(
-        src_dataset,
-        src_vocab_table,
-        batch_size=batch_size_placeholder,
-        eos=hparams.eos,
-        source_reverse=hparams.source_reverse,
-        src_max_len=hparams.src_max_len_infer)
-    model = model_creator(
-        hparams,
-        iterator=iterator,
-        mode=tf.contrib.learn.ModeKeys.INFER,
-        source_vocab_table=src_vocab_table,
-        target_vocab_table=tgt_vocab_table,
-        reverse_target_vocab_table=reverse_tgt_vocab_table,
-        scope=scope,
-        single_cell_fn=single_cell_fn)
-  return InferModel(
-      graph=graph,
-      model=model,
-      src_placeholder=src_placeholder,
-      batch_size_placeholder=batch_size_placeholder,
-      iterator=iterator)
 
 
 def _decode_inference_indices(model, sess, output_infer,
@@ -139,8 +88,7 @@ def inference(ckpt,
               hparams,
               num_workers=1,
               jobid=0,
-              scope=None,
-              single_cell_fn=None):
+              scope=None):
   """Perform translation."""
   if hparams.inference_indices:
     assert num_workers == 1
@@ -153,8 +101,7 @@ def inference(ckpt,
     model_creator = gnmt_model.GNMTModel
   else:
     raise ValueError("Unknown model architecture")
-  infer_model = create_infer_model(model_creator, hparams, scope,
-                                   single_cell_fn)
+  infer_model = model_helper.create_infer_model(model_creator, hparams, scope)
 
   if num_workers == 1:
     single_worker_inference(
